提供模型步骤信息服务

+++ src/Dist.Product.Dme/Dist.Dme.Model/Entity/DmeRuleStep.cs
@@ -16,9 +16,9 @@ namespace Dist.Dme.Model.Entity
         [SugarColumn(ColumnName = "VERSION_ID")]
         public int VersionId { get; set; }
         [SugarColumn(ColumnName = "GUI_LOCATION_X")]
-        public int GuiLocationX { get; set; }
+        public double GuiLocationX { get; set; }
         [SugarColumn(ColumnName = "GUI_LOCATION_Y")]
-        public int GuiLocationY { get; set; }
+        public double GuiLocationY { get; set; }
         [SugarColumn(ColumnName = "STEP_TYPE_ID")]
         public int StepTypeId { get; set; }
     }
